requirements.txt
Flask>=2.3.0
Werkzeug>=2.3.0
bcrypt>=4.0.0
PyYAML>=6.0
ansible>=7.0.0
pywinrm>=0.3.0
requests>=2.31.0

